refactor: extract a value shared between components to a constant
The toggle mode button has width and we want to add a padding with the same value to the input to avoid text getting under the button.

Change in the button width must be reflected in the padding, so we want to change them together. Constant is a perfect solution for that.